+++ core/report/markdown_templates.go
@@ -1,11 +1,11 @@
 package report
 
-const sloTemplate = `# Reliably SLO Report
+const SLOTemplate = `# Reliably SLO Report
 
 Service Level Objectives identify what you should care about on your system. They are what good looks like for the users of your system. If an SLO is underperforming, it will be impacting your users in some way.
 
 <details>
-  <summary>Expand for further information on SLOs with Reliably</summary>
+  <summary>Expand for further SLOs with Reliably</summary>
 
 
   The Reliably CLI allows you to define SLOs for Availability and Latency.
@@ -16,17 +16,17 @@ Service Level Objectives identify what you should care about on your system. The
 </details>
 
 ### Error Budget
-When you define an SLO for your system, you include a target percentage for that SLO. An example target could be 95%. That leaves 5%, which is your error budget for your SLO.
+When you defiine and SLO for you system, you include a target percentage for that SLO. An example target could be 95%. That leaves 5%, that is your error budget for your SLO.
 
 <details>
-  <summary>Expand for further information on Error Bugets with Reliably</summary>
+  <summary>Expand for further inforamtion on Error Bugets with Reliably</summary>
 
 
   When you define an SLO with the Reliably CLI you specify a target percentage for the Availability or Latency for that SLO.
 
 The expectation is over a time window, the responses for the Serice will be within that target percentage.
 
-An example could be 99.5% available over a period of 7 days. The target availability is less than 100% so this leaves a margin for error. This is the Error Budget.
+An example could be 99.5% available over a period of 7 days. The target availability is less than 100% so this leaves a margin for error. This can be considered the Error Budget.
 
 </details>
 
@@ -35,27 +35,39 @@ For more details of an SLO report, see the Reliably documentation on [How the Re
 
 [How the Reliably CLI works]:https://reliably.com/docs/guides/how-it-works/slo-reports/
 
-Report time: {{ dateTime .Rep.Timestamp }}
+{{ $report := .Rep }}
+
+Report time: {{ dateTime $report.Timestamp }}
 {{ $reps := .Lreps }}
-{{ range $index, $service := .Rep.Services }}
+{{ range $index, $service := $report.Services }}
 ## Service #{{ serviceNo $index}}: {{$service.Name}}
 
-|  | Name               | Current | Objective| Time Window | Type  | Trend  |
-|--|------------------- | ------- | -------- |------------ |-------|-------|
-{{ range $ind, $sl := $service.ServiceLevels }}{{ serviceLevelRow $service.Name $sl $reps }}
+|  | Name                            | Current | Objective| Time Window | Type  | Trend |
+|--|-------------------------------- | ------- | -------- |------------ |-------|-------|
+{{ range $ind, $sl := $service.ServiceLevels -}}
+	|{{- svcLevelGetStatusIcon $sl -}}
+	|{{- svcLevelGetName $sl}}|{{svcLevelGetActualResult $sl}}|{{ svcLevelGetObjective $sl }}|{{ svcLevelGetTimeWindow $sl }}|{{ svcLevelGetType $sl }}|{{ svcLevelGetTrends $service.Name $sl $reps }}
 {{ end }}
 
 
 
 The Error Budget metrics are:
 
-|  Type    | Name          |ErrorBudget(%)|Downtime|Consumed|Remain
-| -------- | --------------|--------------|--------|--------|--------|
-{{ range $ind, $sl := $service.ServiceLevels }}{{ errorBudgetRow $sl }}
+|  Type    | Name          |ErrorBudget(%)|Time Window|Downtime|Consumed|Remain
+| -------- | --------------|--------------|-----------|--------|--------|--------|
+{{ range $ind, $sl := $service.ServiceLevels -}}
+|{{- svcLevelGetType $sl -}}|
+{{- svcLevelGetName $sl}}|
+{{- svcLevelGetTimeWindow $sl }}|
+{{- errBudgetPercentage $sl }}|
+{{- errBudgetAllowedDownTime $sl }}|
+{{- errBudgetConsumed $sl }}|
+{{- errBudgetRemain $sl }}|
 {{ end }}
 
 
 {{ end }}
 
-<sub>Generating with: The Reliably CLI Version {{ version }}</sub>
+<sub>Generated with: The Reliably CLI Version {{ reliablyVersion }}</sub>
+
 `