+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+### Added
+
 -   changes SLO report to support reports from templates; [#316](https://github.com/reliablyhq/cli/issues/316)
 
 ## [v0.16.1] - 2021-06-01